[src] Fix some warnings
(cherry picked from commit 2bcc0a7a4bb5c59384acdceead9c9bf1bd9a6db4)

+++ src/InteractionTools/SurfaceCarvingManager.cpp
@@ -184,7 +184,7 @@ void SurfaceCarvingManager::CheckCollisionDetection()
 		{
 			contacts = dynamic_cast<const ContactVector*>(it->second);
 		}
-		unsigned int ncontacts = 0;
+		size_t ncontacts = 0;
 		if (contacts != NULL)
 		{
 			Coll_TF = true;
